+++ src/utils/_utils.d.ts
@@ -2,36 +2,42 @@ declare namespace $ext {
   /** 工作区路径 */
   const WORKSPACE_PATH: string | undefined
   /** 插件设置 */
-  const CONFIG_LIST:
-    | 'globalTemplatePath'
-    | 'workspaceTemplatePath'
-    | 'defaultFolderTemplate'
-    | 'defaultFileTemplate'[]
+  const CONFIG_LIST: Array<keyof CodeConfig>
+  /** 插件发布者 */
+  const PUBLISHER: string
   /** 插件名称 */
   const EXT_NAME: string
   /** 插件私有配置文件路径 */
   const LOCAL_CONFIG_PATH: string
   /** 插件本体路径 */
   const EXT_PATH: string
-  /** 默认缩进单位 */
-  const BASE_INDENTATION: string
-  /** 默认缩进宽度 */
-  const BASE_INDENTATION_COUNT: number
+  /** 模板配置文件名 */
+  const TEMPLATE_CONFIG_FILE_NAME: string
 
   namespace config {
     /** 插件配置 */
     const extConfig: ExtConfig
     /** 获取工作区配置 */
-    function getWorkspaceConfig(): WorkspaceConfig
+    function getWorkspaceConfig(): CodeConfig
     /** 写入工作区配置 */
-    function setWorkspaceConfig(config: WorkspaceConfig): void
+    function setWorkspaceConfig(config: CodeConfig): void
     /** 获取插件目录下的配置 */
     function getLocalConfig(): LocalConfig
     /** 写入插件目录下的配置 */
     function setLocalConfig(config: LocalConfig): void
+    /**
+     * 获取全局配置文件路径
+     * @param fileName 文件名
+     */
+    function getGlobalStoragePath(fileName?: string): string
   }
 
   namespace localize {
+    /**
+     * 多语言转换
+     * @param key
+     * @param params
+     */
     function getLocalize(key: string, ...params: string[]): string
   }
 
@@ -65,18 +71,11 @@ declare namespace $ext {
   function preSaveDocument(docStr: string, name: string): Thenable<boolean>
 }
 
-interface WorkspaceConfig {
-  savePath?: string
-  swaggerJsonUrl?: SwaggerJsonUrlItem[]
+interface CodeConfig {
+  defaultFolderTemplate?: string
+  defaultFileTemplate?: string
 }
 
-interface SwaggerJsonUrlItem {
-  title: string
-  url: string
-}
+interface ExtConfig extends CodeConfig, LocalConfig {}
 
-interface ExtConfig extends WorkspaceConfig, LocalConfig {}
-
-interface LocalConfig {
-  activeGroupIndex?: number
-}
+interface LocalConfig {}