+++ src/utils/config.ts
@@ -1,39 +1,74 @@
-import { workspace } from 'vscode'
+import vscode, { workspace } from 'vscode'
 import fs from 'fs'
+import path from 'path'
 
-import { LOCAL_CONFIG_PATH, EXT_NAME, CONFIG_LIST } from './const'
+import { LOCAL_CONFIG_PATH, CONFIG_LIST, EXT_NAME, PUBLISHER } from './const'
 import log from './log'
+
+const CONFIG_GROUP = 'createItemByTemplate'
 
 class Config {
   /**
    * 获取全部配置
    */
   get extConfig(): ExtConfig {
-    return { ...this.getWorkspaceConfig(), ...this.getLocalConfig() }
+    return { ...this.getCodeConfig(), ...this.getLocalConfig() }
   }
 
   /**
-   * 获取工作区配置
+   * 获取 vscode 配置
    */
-  getWorkspaceConfig(): WorkspaceConfig {
+  getCodeConfig(): CodeConfig {
     const resConfig = {}
     CONFIG_LIST.forEach(configKey => {
-      const settingsKey = `${EXT_NAME}.${configKey}`
+      const settingsKey = `${CONFIG_GROUP}.${configKey}`
       resConfig[configKey] = workspace.getConfiguration().get(settingsKey)
     })
     return resConfig
   }
 
-  setWorkspaceConfig(config: WorkspaceConfig) {
+  /**
+   * 写入 vscode 配置
+   * @param config
+   */
+  setCodeConfig(config: CodeConfig) {
     for (const configKey in config) {
-      const settingsKey = `${EXT_NAME}.${configKey}`
+      const settingsKey = `${CONFIG_GROUP}.${configKey}`
       const val = config[configKey]
       workspace.getConfiguration().update(settingsKey, val, false)
     }
   }
 
+  getChannelPath() {
+    if (vscode.env.appName.indexOf('Insiders') > 0) {
+      return 'Code - Insiders'
+    } else {
+      return 'Code'
+    }
+  }
+
   /**
-   * 获取本地配置
+   * 获取全局配置文件路径
+   * @param fileName 文件名
+   */
+  getGlobalStoragePath(fileName = ''): string {
+    let appdata =
+      process.env.APPDATA ||
+      (process.platform === 'darwin' ? process.env.HOME + '/Library/Application Support' : '/var/local')
+    let channelPath = this.getChannelPath()
+    const globalStoragePath = path.join(
+      appdata,
+      channelPath,
+      'User',
+      'globalStorage',
+      `${PUBLISHER}.${EXT_NAME}`,
+      fileName
+    )
+    return globalStoragePath
+  }
+
+  /**
+   * 获取本地文件配置
    */
   getLocalConfig(): LocalConfig {
     let config = {}
@@ -51,6 +86,10 @@ class Config {
     return config
   }
 
+  /**
+   * 写入本地文件配置
+   * @param config
+   */
   setLocalConfig(config: LocalConfig) {
     const defaultConfig = this.getLocalConfig()
     try {