+++ src/utils/io.ts
@@ -3,7 +3,6 @@ import path from 'path'
 import vscode from 'vscode'
 
 import { WORKSPACE_PATH } from './const'
-// import config from './config'
 
 /**
  * 递归创建路径
@@ -33,8 +32,7 @@ export function preSaveDocument(docStr: string, filePath: string): Thenable<bool
   return vscode.workspace.openTextDocument(newFile).then(document => {
     const edit = new vscode.WorkspaceEdit()
     const pMin = new vscode.Position(0, 0)
-    const pMax = new vscode.Position(100000000, 100000000)
-    // edit.insert(newFile, pMin, 'Hello world!' + JSON.stringify(e))
+    const pMax = new vscode.Position(Infinity, Infinity)
     edit.replace(newFile, new vscode.Range(pMin, pMax), docStr)
     return vscode.workspace.applyEdit(edit).then(success => {
       if (success) {